EES-2520 Merge data API's ThemeService with content API
This implements the previous endpoint as a new `publicationFilter`
query parameter for the content API's `/themes` endpoint.
