+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services.Tests/ThemeServiceTests.cs
@@ -1,12 +1,14 @@
 #nullable enable
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Tests.Utils;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using Moq;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
@@ -723,6 +725,351 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Tests
 
                 Assert.Equal("Publication C", publications[1].Title);
                 Assert.Equal("https://legacy.url/", publications[1].LegacyPublicationUrl);
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData()
+        {
+            var releaseFile1 = new ReleaseFile
+            {
+                Release = new Release
+                {
+                    ReleaseName = "2000",
+                    Published = DateTime.UtcNow,
+                    Publication = new Publication
+                    {
+                        Title = "Theme 1 topic 1 publication 1",
+                        Slug = "theme-a-topic-a-publication-1",
+                        Topic = new Topic
+                        {
+                            Title = "Theme 1 topic 1",
+                            Slug = "theme-a-topic-a",
+                            Theme = new Theme
+                            {
+                                Title = "Theme 1",
+                                Slug = "theme-a",
+                            }
+                        }
+                    }
+                },
+                File = new File
+                {
+                    Type = FileType.Data
+                }
+            };
+
+            var releaseFile2 = new ReleaseFile
+            {
+                Release = new Release
+                {
+                    ReleaseName = "2000",
+                    Published = DateTime.UtcNow,
+                    Publication = new Publication
+                    {
+                        Title = "Theme B topic B publication B",
+                        Slug = "theme-b-topic-b-publication-B",
+                        Topic = new Topic
+                        {
+                            Title = "Theme B topic B",
+                            Slug = "theme-b-topic-b",
+                            Theme = new Theme
+                            {
+                                Title = "Theme B",
+                                Slug = "theme-b",
+                            }
+                        }
+                    }
+                },
+                File = new File
+                {
+                    Type = FileType.Data
+                }
+            };
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                await context.AddRangeAsync(releaseFile1, releaseFile2);
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+
+                var result = await service.GetPublicationTree(PublicationTreeFilter.LatestData);
+
+                Assert.Equal(2, result.Count);
+                Assert.Equal(releaseFile1.Release.Publication.Topic.Theme.Id, result[0].Id);
+                Assert.Equal(releaseFile1.Release.Publication.Topic.Theme.Title, result[0].Title);
+
+                var themeATopics = result[0].Topics.ToList();
+                Assert.Single(themeATopics);
+                Assert.Equal(releaseFile1.Release.Publication.Topic.Id, themeATopics[0].Id);
+                Assert.Equal(releaseFile1.Release.Publication.Topic.Title, themeATopics[0].Title);
+
+                var themeATopicAPublications = themeATopics[0].Publications.ToList();
+                Assert.Single(themeATopicAPublications);
+                Assert.Equal(releaseFile1.Release.Publication.Id, themeATopicAPublications[0].Id);
+                Assert.Equal(releaseFile1.Release.Publication.Slug, themeATopicAPublications[0].Slug);
+                Assert.Equal(releaseFile1.Release.Publication.Title, themeATopicAPublications[0].Title);
+
+                Assert.Equal(releaseFile2.Release.Publication.Topic.Theme.Id, result[1].Id);
+                Assert.Equal(releaseFile2.Release.Publication.Topic.Theme.Title, result[1].Title);
+
+                var themeBTopics = result[1].Topics.ToList();
+                Assert.Single(themeBTopics);
+                Assert.Equal(releaseFile2.Release.Publication.Topic.Id, themeBTopics[0].Id);
+                Assert.Equal(releaseFile2.Release.Publication.Topic.Title, themeBTopics[0].Title);
+
+                var themeBTopicAPublications = themeBTopics[0].Publications.ToList();
+                Assert.Single(themeBTopicAPublications);
+                Assert.Equal(releaseFile2.Release.Publication.Id, themeBTopicAPublications[0].Id);
+                Assert.Equal(releaseFile2.Release.Publication.Slug, themeBTopicAPublications[0].Slug);
+                Assert.Equal(releaseFile2.Release.Publication.Title, themeBTopicAPublications[0].Title);
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData_ThemeHasNoTopics()
+        {
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var theme = new Theme
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Theme",
+                    Slug = "theme"
+                };
+
+                await context.AddAsync(theme);
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+                Assert.Empty(await service.GetPublicationTree(PublicationTreeFilter.LatestData));
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData_TopicHasNoPublications()
+        {
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var theme = new Theme
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Theme",
+                    Slug = "theme"
+                };
+
+                var topic = new Topic
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Topic",
+                    Slug = "topic",
+                    ThemeId = theme.Id
+                };
+
+                await context.AddAsync(theme);
+                await context.AddAsync(topic);
+
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+                Assert.Empty(await service.GetPublicationTree(PublicationTreeFilter.LatestData));
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData_PublicationHasNoReleases()
+        {
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var theme = new Theme
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Theme",
+                    Slug = "theme"
+                };
+
+                var topic = new Topic
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Topic",
+                    Slug = "topic",
+                    ThemeId = theme.Id
+                };
+
+                var publication = new Publication
+                {
+                    Id = Guid.NewGuid(),
+                    Title = "Publication",
+                    Slug = "publication",
+                    TopicId = topic.Id
+                };
+
+                await context.AddAsync(theme);
+                await context.AddAsync(topic);
+                await context.AddAsync(publication);
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+                Assert.Empty(await service.GetPublicationTree(PublicationTreeFilter.LatestData));
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData_PublicationHasNoPublishedReleases()
+        {
+            var theme = new Theme
+            {
+                Title = "Theme",
+                Slug = "theme"
+            };
+
+            var topic = new Topic
+            {
+                Title = "Topic",
+                Slug = "topic",
+                Theme = theme
+            };
+
+            var publicationA = new Publication
+            {
+                Title = "Publication A",
+                Slug = "publication-a",
+                Topic = topic
+            };
+
+            var publicationB = new Publication
+            {
+                Title = "Publication B",
+                Slug = "publication-b",
+                Topic = topic
+            };
+
+            var publicationARelease = new Release
+            {
+                Publication = publicationA,
+                ReleaseName = "2000",
+                Published = DateTime.UtcNow
+            };
+
+            var publicationBRelease = new Release
+            {
+                Publication = publicationB,
+                ReleaseName = "2000",
+                Published = null
+            };
+
+            var publicationAReleaseFile = new ReleaseFile
+            {
+                Release = publicationARelease,
+                File = new File
+                {
+                    Type = FileType.Data
+                }
+            };
+
+            var publicationBReleaseFile = new ReleaseFile
+            {
+                Release = publicationBRelease,
+                File = new File
+                {
+                    Type = FileType.Data
+                }
+            };
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                await context.AddRangeAsync(publicationAReleaseFile, publicationBReleaseFile);
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+
+                var result = await service.GetPublicationTree(PublicationTreeFilter.LatestData);
+
+                Assert.Single(result);
+                Assert.Equal(theme.Id, result[0].Id);
+                Assert.Equal(theme.Title, result[0].Title);
+                var topics = result[0].Topics.ToList();
+
+                Assert.Single(topics);
+                Assert.Equal(topic.Id, topics[0].Id);
+                Assert.Equal(topic.Title, topics[0].Title);
+                var publications = topics[0].Publications.ToList();
+
+                Assert.Single(publications);
+                Assert.Equal(publicationA.Id, publications[0].Id);
+                Assert.Equal(publicationA.Slug, publications[0].Slug);
+                Assert.Equal(publicationA.Title, publications[0].Title);
+
+                // Publication B is not included because it's Release is not published
+            }
+        }
+
+        [Fact]
+        public async Task GetPublicationTree_LatestData_PublicationLatestReleaseHasNoSubject()
+        {
+            var theme = new Theme();
+            var topic = new Topic { Theme = theme };
+            var publication = new Publication { Topic = topic };
+            var previousRelease = new Release
+            {
+                Publication = publication,
+                ReleaseName = "2000",
+                Published = DateTime.UtcNow.Subtract(TimeSpan.FromDays(1))
+            };
+            var previousReleaseFile = new ReleaseFile
+            {
+                Release = previousRelease,
+                File = new File { Type = FileType.Data }
+            };
+
+            var latestRelease = new Release
+            {
+                Publication = publication,
+                ReleaseName = "2000",
+                Published = DateTime.UtcNow,
+                PreviousVersion = previousRelease
+            };
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                await context.AddRangeAsync(previousReleaseFile, latestRelease);
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = BuildThemeService(context);
+
+                var result = await service.GetPublicationTree(PublicationTreeFilter.LatestData);
+                Assert.Empty(result);
             }
         }
