+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/Interfaces/IThemeService.cs
@@ -1,13 +1,14 @@
 #nullable enable
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 
 namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces
 {
     public interface IThemeService
     {
-        Task<IList<ThemeTree<PublicationTreeNode>>> GetPublicationTree();
+        Task<IList<ThemeTree<PublicationTreeNode>>> GetPublicationTree(PublicationTreeFilter? filter = null);
 
         Task<IList<ThemeTree<PublicationDownloadsTreeNode>>> GetPublicationDownloadsTree();
     }
