+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/Requests/PublicationTreeFilter.cs
@@ -1,0 +1,9 @@
+#nullable enable
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Requests
+{
+    public enum PublicationTreeFilter
+    {
+        LatestData
+    }
+}
