+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ThemeService.cs
@@ -4,11 +4,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Cache;
-using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Microsoft.EntityFrameworkCore;
 
@@ -26,51 +27,81 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
         }
 
         [BlobCache(typeof(PublicationTreeCacheKey))]
-        public async Task<IList<ThemeTree<PublicationTreeNode>>> GetPublicationTree()
+        public async Task<IList<ThemeTree<PublicationTreeNode>>> GetPublicationTree(
+            PublicationTreeFilter? filter = null)
         {
-            var themes = await _contentDbContext.Themes
-                .Include(theme => theme.Topics)
-                .ThenInclude(topic => topic.Publications)
-                .ThenInclude(publication => publication.Releases)
+            var themes = await ListThemes();
+
+            return await themes
+                .ToAsyncEnumerable()
+                .SelectAwait(async theme => await BuildThemeTree(theme, filter))
+                .Where(theme => theme.Topics.Any())
+                .OrderBy(theme => theme.Title)
+                .ToListAsync();
+        }
+
+        private async Task<ThemeTree<PublicationTreeNode>> BuildThemeTree(
+            Theme theme,
+            PublicationTreeFilter? filter)
+        {
+            var topics = await theme.Topics
+                .ToAsyncEnumerable()
+                .SelectAwait(async topic => await BuildTopicTree(topic, filter))
+                .Where(topic => topic.Publications.Any())
+                .OrderBy(topic => topic.Title)
                 .ToListAsync();
 
-            return themes
-                .Where(theme => IsThemePublished(theme, IsPublicationPublishedOrHasLegacyUrl))
-                .Select(
-                    theme => new ThemeTree<PublicationTreeNode>()
-                    {
-                        Id = theme.Id,
-                        Title = theme.Title,
-                        Summary = theme.Summary,
-                        Topics = theme.Topics
-                            .Where(topic => IsTopicPublished(topic, IsPublicationPublishedOrHasLegacyUrl))
-                            .Select(BuildTopicTree)
-                            .OrderBy(topic => topic.Title)
-                            .ToList()
-                    }
-                )
-                .OrderBy(theme => theme.Title)
-                .ToList();
+            return new ThemeTree<PublicationTreeNode>
+            {
+                Id = theme.Id,
+                Title = theme.Title,
+                Summary = theme.Summary,
+                Topics = topics
+            };
         }
 
-        private static bool IsPublicationPublishedOrHasLegacyUrl(Publication publication)
+        private async Task<TopicTree<PublicationTreeNode>> BuildTopicTree(
+            Topic topic,
+            PublicationTreeFilter? filter)
         {
-            return !string.IsNullOrEmpty(publication.LegacyPublicationUrl?.ToString())
-                   || publication.Releases.Any(release => release.IsLatestPublishedVersionOfRelease());
-        }
+            var publications = await topic.Publications
+                .ToAsyncEnumerable()
+                .WhereAwait(async publication => await FilterPublication(publication, filter))
+                .Select(BuildPublicationNode)
+                .OrderBy(publication => publication.Title)
+                .ToListAsync();
 
-        private static TopicTree<PublicationTreeNode> BuildTopicTree(Topic topic)
-        {
             return new()
             {
                 Id = topic.Id,
                 Title = topic.Title,
-                Publications = topic.Publications
-                    .Where(IsPublicationPublishedOrHasLegacyUrl)
-                    .Select(BuildPublicationNode)
-                    .OrderBy(publication => publication.Title)
-                    .ToList()
+                Publications = publications
             };
+        }
+
+        private async Task<bool> FilterPublication(
+            Publication publication,
+            PublicationTreeFilter? filter)
+        {
+            switch (filter)
+            {
+                case PublicationTreeFilter.LatestData:
+                    var latestLiveRelease = publication.LatestPublishedRelease();
+
+                    return latestLiveRelease != null && await HasAnyDataFiles(latestLiveRelease);
+                case null:
+                    return !string.IsNullOrEmpty(publication.LegacyPublicationUrl?.ToString()) ||
+                           publication.Releases.Any(release => release.IsLatestPublishedVersionOfRelease());
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(filter), filter, null);
+            }
+        }
+
+        private async Task<bool> HasAnyDataFiles(Release release)
+        {
+            return await _contentDbContext.ReleaseFiles
+                .Include(rf => rf.File)
+                .AnyAsync(rf => rf.ReleaseId == release.Id && rf.File.Type == FileType.Data);
         }
 
         private static PublicationTreeNode BuildPublicationNode(Publication publication)
@@ -78,8 +109,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
             // Ignore any legacyPublicationUrl once the Publication has Releases
             var legacyPublicationUrlIgnored =
                 publication.Releases.Any(release => release.IsLatestPublishedVersionOfRelease());
-            var legacyPublicationUrl =
-                legacyPublicationUrlIgnored ? null : publication.LegacyPublicationUrl?.ToString();
 
             return new PublicationTreeNode
             {
@@ -87,51 +116,52 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
                 Title = publication.Title,
                 Summary = publication.Summary,
                 Slug = publication.Slug,
-                LegacyPublicationUrl = legacyPublicationUrl
+                LegacyPublicationUrl = legacyPublicationUrlIgnored
+                    ? null
+                    : publication.LegacyPublicationUrl?.ToString()
             };
         }
 
+        // TODO: EES-2365 Remove once 'Download latest data' page no longer exists
         [BlobCache(typeof(PublicationDownloadsTreeCacheKey))]
         public async Task<IList<ThemeTree<PublicationDownloadsTreeNode>>> GetPublicationDownloadsTree()
         {
-            var themes = await _contentDbContext.Themes
-                .Include(theme => theme.Topics)
-                .ThenInclude(topic => topic.Publications)
-                .ThenInclude(publication => publication.Releases)
-                .ToListAsync();
+            var themes = await ListThemes();
 
-            var trees = await themes
-                .Where(theme => IsThemePublished(theme))
-                .SelectAsync(
-                    async theme =>
-                    {
-                        var topics = await theme.Topics
-                            .Where(topic => IsTopicPublished(topic))
-                            .SelectAsync(async topic => await BuildTopicDownloadsTree(topic));
-
-                        return new ThemeTree<PublicationDownloadsTreeNode>
-                        {
-                            Id = theme.Id,
-                            Title = theme.Title,
-                            Summary = theme.Summary,
-                            Topics = topics
-                                .Where(topic => topic.Publications.Any())
-                                .OrderBy(topic => topic.Title).ToList()
-                        };
-                    }
-                );
-
-            return trees
+            return await themes
+                .ToAsyncEnumerable()
+                .SelectAwait(async theme => await BuildThemeDownloadsTree(theme))
                 .Where(theme => theme.Topics.Any())
                 .OrderBy(theme => theme.Title)
-                .ToList();
+                .ToListAsync();
+        }
+
+        private async Task<ThemeTree<PublicationDownloadsTreeNode>> BuildThemeDownloadsTree(Theme theme)
+        {
+            var topics = await theme.Topics
+                .ToAsyncEnumerable()
+                .SelectAwait(async topic => await BuildTopicDownloadsTree(topic))
+                .Where(topic => topic.Publications.Any())
+                .ToListAsync();
+
+            return new ThemeTree<PublicationDownloadsTreeNode>
+            {
+                Id = theme.Id,
+                Title = theme.Title,
+                Summary = theme.Summary,
+                Topics = topics
+                    .Where(topic => topic.Publications.Any())
+                    .OrderBy(topic => topic.Title).ToList()
+            };
         }
 
         private async Task<TopicTree<PublicationDownloadsTreeNode>> BuildTopicDownloadsTree(Topic topic)
         {
             var publications = await topic.Publications
-                .Where(publication => IsPublicationPublished(publication))
-                .SelectAsync(async publication => await BuildPublicationDownloadsNode(publication));
+                .ToAsyncEnumerable()
+                .Where(IsPublicationPublished)
+                .SelectAwait(async publication => await BuildPublicationDownloadsNode(publication))
+                .ToListAsync();
 
             return new TopicTree<PublicationDownloadsTreeNode>
             {
@@ -167,20 +197,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
             };
         }
 
-        private static bool IsThemePublished(Theme theme, Func<Publication, bool>? publicationCriteria = null)
+        private async Task<List<Theme>> ListThemes()
         {
-            return theme.Topics.Any(topic => IsTopicPublished(topic, publicationCriteria));
+            return await _contentDbContext.Themes
+                .Include(theme => theme.Topics)
+                .ThenInclude(topic => topic.Publications)
+                .ThenInclude(publication => publication.Releases)
+                .ToListAsync();
         }
 
-        private static bool IsTopicPublished(Topic topic, Func<Publication, bool>? publicationCriteria = null)
+        private static bool IsPublicationPublished(Publication publication)
         {
-            return topic.Publications.Any(publication => IsPublicationPublished(publication, publicationCriteria));
-        }
-
-        private static bool IsPublicationPublished(Publication publication, Func<Publication, bool>? criteria = null)
-        {
-            return criteria?.Invoke(publication) ??
-                   publication.Releases.Any(release => release.IsLatestPublishedVersionOfRelease());
+            return publication.Releases.Any(release => release.IsLatestPublishedVersionOfRelease());
         }
     }
 }
