--- src/GovUk.Education.ExploreEducationStatistics.Data.Api/Controllers/ThemeController.cs
@@ -1,26 +1,0 @@
-using System.Collections.Generic;
-using System.Linq;
-using GovUk.Education.ExploreEducationStatistics.Data.Services.Interfaces;
-using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
-using Microsoft.AspNetCore.Mvc;
-
-namespace GovUk.Education.ExploreEducationStatistics.Data.Api.Controllers
-{
-    [Route("api/themes")]
-    [ApiController]
-    public class ThemeController : ControllerBase
-    {
-        private readonly IThemeService _themeService;
-
-        public ThemeController(IThemeService themeService)
-        {
-            _themeService = themeService;
-        }
-
-        [HttpGet]
-        public ActionResult<IEnumerable<ThemeViewModel>> ListThemesWithLiveSubjects()
-        {
-            return _themeService.ListThemesWithLiveSubjects().ToList();
-        }
-    }
-}
