+++ src/explore-education-statistics-common/src/modules/table-tool/components/PublicationForm.tsx
@@ -8,7 +8,7 @@ import {
 } from '@common/components/form';
 import SummaryList from '@common/components/SummaryList';
 import SummaryListItem from '@common/components/SummaryListItem';
-import { Theme } from '@common/services/tableBuilderService';
+import { Theme } from '@common/services/themeService';
 import createErrorHelper from '@common/validation/createErrorHelper';
 import Yup from '@common/validation/yup';
 import { Formik } from 'formik';
