+++ src/explore-education-statistics-common/src/modules/table-tool/components/__tests__/TableToolWizard.test.tsx
@@ -1,9 +1,6 @@
 import TableToolWizard from '@common/modules/table-tool/components/TableToolWizard';
-import {
-  Subject,
-  SubjectMeta,
-  Theme,
-} from '@common/services/tableBuilderService';
+import { Subject, SubjectMeta } from '@common/services/tableBuilderService';
+import { Theme } from '@common/services/themeService';
 import { within } from '@testing-library/dom';
 import { render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
@@ -13,17 +10,18 @@ describe('TableToolWizard', () => {
     {
       id: 'theme-1',
       title: 'Theme 1',
-      slug: 'theme-1',
+      summary: '',
       topics: [
         {
           id: 'topic-1',
           title: 'Topic 1',
-          slug: 'topic-1',
+          summary: '',
           publications: [
             {
               id: 'publication-1',
               title: 'Publication 1',
               slug: 'publication-1',
+              summary: '',
             },
           ],
         },
