+++ src/explore-education-statistics-frontend/src/modules/data-catalogue/__tests__/DataCataloguePage.test.tsx
@@ -19,7 +19,6 @@ describe('DataCataloguePage', () => {
         {
           id: 'topic-1',
           title: 'Admission appeals',
-          slug: 'admission-appeals',
           publications: [
             {
               id: 'publication-1',
